isEssentiallyCNF, pushWeakeningsUp: move to appropriate packages
